backend/backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
python-dotenv
psycopg2-binary //failed install